Moved placeholder tutorial images + Credits Scene
Moved the placeholder images into the "placeholder sprites" folder.
Added a new scene "Credits"
Wrote the credits to the game in the Credits scene.
Edited Main Menu to include "Credits".
Edited ManiMenu object to include a "Credits" button.
Edited MainMenuScript to include Credits and ReturnToMenu scene transitions.

Known issues: Level does not load because I cannot edit what levels are build when starting the game. Need to add the Credits scene as scene #4 for it to work properly